transform to dataframe, remove the denpendency of pycall and lost some functions
