+++ src/EMIRT.jl
@@ -8,7 +8,7 @@ include("core/boundary.jl")
 #include("emshow.jl")
 include("core/emio.jl")
 include("core/label.jl")
-include("core/evaluate.jl")
+#include("core/evaluate.jl")
 include("core/emparser.jl")
 include("core/filesystem.jl")
 include("core/utils.jl")
