+++ src/plugins/emshow.jl
@@ -1,11 +1,11 @@
 export random_color_show, plot
 using PyPlot
 using Gadfly
-using Vega
 
 include("../core/label.jl")
 include("../core/evaluate.jl")
 include("../core/errorcurve.jl")
+include("utils.jl")
 
 function random_color_show( seg )
     seg2 = seg
@@ -26,33 +26,24 @@ function random_color_show( seg )
 end
 
 """
-plot the error curve
-"""
-function plot(ec::Tec, clr="blue")
-    prf = plot(layer(x=ec["thd"], y=ec["rf"],
-                     Geom.line, Theme(default_color=color(clr))),
-               Guide.xlabel("threshold"), Guide.ylabel("rand f score")))
-    pre = plot(layer(x=ec["thd"], y=ec["re"],
-                     Geom.line, Theme(default_color=color(clr))),
-    Guide.xlabel("threshold"), Guide.ylabel("rand error")))
-
-    prfms = plot(layer(x=ec["rfm"], y=ec["rfs"],
-    Geom.line, Theme(default_color=color(clr))),
-                 Guide.xlabel("rand f score of merging"), Guide.ylabel("rand f score of splitting")))
-    prems = plot(layer(x=ec["rem"], y=ec["res"],
-                       Geom.line, Theme(default_color=color(clr))),
-                 Guide.xlabel("rand error of mergers"), Guide.ylabel("rand error of splitters")))
-    # stack the subplots
-    plt = vstack(hstack(prf, prfms), hstack(pre, prems))
-end
-
-"""
 plot multiple error curves
 """
 function plot(ecs::Tecs)
-    thds = Vector{Vector{Float32}}()
-    rf = 
-    for tag, ec in Tecs
-        continue
-    end
+    # transform to dataframe
+    df = ecs2df(ecs)
+    # plot the dataframe
+    prf = plot(df, x="thd", y="rf", color="tag", Geom.line,
+               Guide.xlabel("threshold"),
+               Guide.ylabel("rand f score"))
+    pre = plot(df, x="thd", y="re", color="tag", Geom.line,
+               Guide.xlabel("threshold"),
+               Guide.ylabel("rand error"))
+    prfms = plot(df, x="rfm", y="rfs", color="tag", Geom.line,
+                 Guide.xlabel("rand f score of mergers"),
+                 Guide.ylabel("rand f score of splitters"))
+    prems = plot(df, x="rem", y="res", color="tag", Geom.line,
+                 Guide.xlabel("rand error of mergers"),
+                 Guide.ylabel("rand error of splitters"))
+    # stack the subplots
+    plt = vstack(hstack(prf,prfms), hstack(pre, prems))
 end
